+++ services/cronjob/nodecleaner/cmd/nodecleaner/main.go
@@ -3,15 +3,13 @@ package main
 import (
 	"github.com/MurmurationsNetwork/MurmurationsServices/common/logger"
 	"github.com/MurmurationsNetwork/MurmurationsServices/common/mongo"
-	"github.com/MurmurationsNetwork/MurmurationsServices/services/cronjob/nodecleaner/config"
-	"github.com/MurmurationsNetwork/MurmurationsServices/services/cronjob/nodecleaner/internal/adapter/mongodb"
+	"github.com/MurmurationsNetwork/MurmurationsServices/services/cronjob/nodecleaner/global"
 	"github.com/MurmurationsNetwork/MurmurationsServices/services/cronjob/nodecleaner/internal/repository/db"
 	"github.com/MurmurationsNetwork/MurmurationsServices/services/cronjob/nodecleaner/internal/service"
 )
 
 func init() {
-	config.Init()
-	mongodb.Init()
+	global.Init()
 }
 
 func main() {
