+++ services/validation/internal/adapter/controller/http/ping_handler.go
