+++ services/validation/internal/app/application.go
@@ -7,13 +7,12 @@ import (
 	"github.com/MurmurationsNetwork/MurmurationsServices/common/logger"
 	"github.com/MurmurationsNetwork/MurmurationsServices/common/middleware"
 	"github.com/MurmurationsNetwork/MurmurationsServices/services/validation/config"
-	"github.com/MurmurationsNetwork/MurmurationsServices/services/validation/internal/adapter/nats"
+	"github.com/MurmurationsNetwork/MurmurationsServices/services/validation/global"
 	"github.com/gin-gonic/gin"
 )
 
 func init() {
-	config.Init()
-	nats.Init()
+	global.Init()
 }
 
 func StartApplication() {
