+++ services/validation/internal/app/url_mapping.go
@@ -1,7 +1,7 @@
 package app
 
 import (
-	"github.com/MurmurationsNetwork/MurmurationsServices/services/validation/internal/controller/http"
+	"github.com/MurmurationsNetwork/MurmurationsServices/services/validation/internal/adapter/controller/http"
 	"github.com/gin-gonic/gin"
 )
