+++ services/validation/internal/entity/node.go
@@ -1,4 +1,4 @@
-package node
+package entity
 
 type Node struct {
 	ProfileURL string `json:"profile_url"`
